SetupBackgroundTasks: Don't allow dialogs during background function execution
Since forever we did not set the dialogsOK parameter for the background
tasks. And the default is dialogsOK = 1 which means that dialogs can run
during background execution.

The Igor help is pretty clear on that, that this can cause crashes when
removing traces from a graph etc.

And that can happen as we call SWS_SaveAcquiredData from the DAQ
background functions and user analysis function as well.

So let's disable that.

Found with user uploaded crash dumps and help from Adam Light
(WaveMetrics).