PA plot: Enhance failed pulse searching
We now allow to search for failed pulses requiring a given number of
pulses. Only if that number of pulses is matched the pulse has passed.

For backwards compatibility, a value of NaN in the GUI means every
non-zero number of pulses is passing.

The algorithm for determining if a pulse has passed or not is also
changed from pure Min/Max to FindLevels.

In addition we now also write the spike positions in the wave note.