WaveMinAndMax: Rename and extend it
In a future commit we want to use the x1, x2 range specifications which
WaveMinMax from IP9 supports.

But for IP procedure functions we have to use named optional arguments,
whereas for builtins these are unnamed. Therefore we have to define
WaveMinAndMax for all IP versions. This also requires to rename it so
that it does not collide with the builtin from IP9.