ReachTargetVoltage: Make it work when using new stimsets with it
In 85d6c5b7 (ReachTargetVoltage: Convert to V3 analysis function, 2020-11-20) we converted that function to the V3 interface.

But this missed the fact that with V3 every analysis function must
handle all events graciously. And that function did not do that.

This was not realized during testing with an existing stimset as
ReachTargetVoltage was still attached to specific events, as always done
for V2 functions, and therefore only called for the supported events
PRE_DAQ_EVENT and POST_SWEEP_EVENT.