DAP_CheckAnalysisFunctionAndParameter: Avoid allowing bogus analysis functions setups
For pre V3 analysis functions we allowed the user to choose the events
to which the analysis function is attached. This is also stored in
separate WPT rows.

Now when you upgrade the analysis function to V3, old stimset still have
it attached only to the previously selected events. These stimsets work
and can be used. But as soon as you try to reattach the analysis
function to the stimset, it will be attached to the generic event. And
then the analysis function will be called for all possible events.

This is quite confusing. So we prevent DAQ in this case and tell the
user how to fix it.