PA plot: Gather more pulse infos
Up to now we only fetch the pulse start times either via the epoch
information from the labnotebook or find it out via FindLevel on the DA
wave.

But having only the pulse starting times lacks the information about the
length of the pulse itself and the length of the baseline.

We now gather all that information and write the pulse high times as
NOTE_KEY_PULSE_START/NOTE_KEY_PULSE_END into the pulse wave note as
well.

This also fixes the issue that the pulse length of single pulses
defaults to the override length, as we could not find out the pulse
length in this case as we required at least two pulses.

Close #732.