PA_PulseHasFailed: Enhance spike position calculation
We actually don't want the level crossing as spike position as it was
done in 8ce7bd07 (PA plot: Enhance failed pulse searching, 2021-01-14).

Better is to use FindPeak which does smoothing and curve discussion to
find the peak location.

Due to that change in the algorithm the PA tests need to be adapted as
well.