tools/unit-testing/RunAllTests.pxp : Fix test failure in Wavebuilder regression
Starting with IP9 37086 the test WB_RegressionTest fails with the
Ref4_a_DA stimset, which has sawtooth epochs.

The failing part of CHECK_EQUAL_WAVES is WAVE_DATA at three points.

These three points are always on the boundary between maximum amplitude,
here 1, and zero.

So before that IP9 version we had

[...]
0.99949998
0
0.00050000002
[...]

and after that we have

[...]
0.99949998
1
0.00050000002
[...]

This looks like a compiler optimization change and does not justify
implementing workarounds on our side.

So let's just branch on the IP version here and use the appropriate
reference waves.