PA_GetPulseInfos: Restructure and return epoch infos if it exists during AUTOMATED_TESTING
During CI tests, when AUTOMATED_TESTING is defined, we used to always
return the calculated pulse infos.

But we also want to return the epoch infos during testing, as that is
what the user will be using on new data.

So let's change that.

And while we are at it, restructure the code and factor PA_DiffPulseInfos out.