ED_createTextNotes: Store timestamps with enough resolution
Since 22a7af4a (ED_createTextNotes: Store sub-second precision in
Timestamp columns, 2017-10-12) we use num2str instead of num2istr for
storing the DateTime in the textual labnotebook.

The intentation was to store timestamps with sub-second precision.

But that change was wrong as Igor outputs only 5 relevant digits for
double when converted to a string as shown in the following example:

•print/D num2str(Datetime)
  3.6968e+09
•print/D num2istr(Datetime)
  3696770484

So we need to use num2strHighPrec with precision 3 which gives
timestamps with millisecond resolution.