Only compile the background watcher panel when required
The background watcher panel is a debugging tool for background tasks.
And therefore by the average user not used at all.

And as it is living in an independent module there is always the chance
that you edit the wrong include file when developing.

So let's not include it by default but only on request.

This also makes compilation faster.