tools/unit-testing/autorun-test.sh: Allow experiment to not exist
Due to our aged bamboo infrastructure for CI we have the same
configuration on all branches. In 7fac38c8 (Tests: Initial PulseAveraging Tests, 2020-11-04) in main a new experiment for testing was added. But that experiment does not exist in the release/2.1 branch.

So let's ignore non-existing experiments to make the tests passing
again.