MIES_AnalysisFunctions.ipf: Clarify post set event speciality
The comment about indexing was introduced in 407fec51 (Analysis
functions: Add PRE_SET_EVENT and fix POST_SET_EVENT for indexing,
2018-06-11).

POST_SET_EVENT works completely with locked and unlocked indexing, the
only speciality is that not all sweeps of a stimset might be acquired.
And only in *that* case is the POST_SET_EVENT not reached.