Hardware Tests: Set interactive mode correctly
This was incorrectly added in c34625b2 (Hardware Tests: Convert to use the new UTF reentry functions, 2019-04-11). The problem is that we set the interactive mode in TEST_BEGIN_OVERRIDE but remove the mies folder in TEST_CASE_BEGIN_OVERRIDE.

So we need to store it in root, and restore, like we do for the
miesVersion.