PA plot: Get the correct graph with non-default region slider
In the PA plot we support looking at the region slider to adjust the
displayed regions.

But if the region slider is active we skip over some regions. And for
finding the first graph we then must not compare not against zero, but the
first region we did not skip.