WB_CalculateParameterWithDelta: Warn on too long explicit delta value list
As we can now access the total number of sweeps, we can warn if we have
too many values.

This will only work if we have at least two sweeps, as WB_AddDelta is
not called for a stimulus set with one sweep as we don't have to do
anything.