CreateTiledChannelGraph: Fix displaying TTL data with only "TP during DAQ" data
Since a2c514a5 (CreateTiledChannelGraph: Move TTL channel display to the
bottom, 2019-03-26) we display TTL channels always at the bottom of the
graph. This is done by checking that we already displayed all AD and DA
channels.

But since 1985c1be (CreateTiledChannelGraph: Don't display data from TP
during DAQ channels, 2020-08-09) we ignore "TP during DAQ" channels. But
in this case we forgot to increase activeChanCount as the TTL displaying
code relies on that.