DC_PlaceDataInDAQDataWave: Fix "TP during DAQ" with TTL channels
We always have two different testpulse length in points units. One is
for TP mode where TTL channels are ignored and one is for DAQ mode where
TTL channels are used.

In DAQ mode we have the inserted TP, but also "TP during DAQ" channels.
And in both cases we have to use the test pulse length which honours TLL
channels.

This is especially difficult to understand in the current code as
testPulseLength is for DAQ mode and TPLength is used whatever is
currently written into.

Fix that by creating the testpulse for the correct mode, so that we can
use it without decimation and drop TPLength.