DC_PlaceDataInDAQDataWave: Fix testpulse zeroing speedup
Since 878eeaf9 (DC_PlaceDataInDAQDataWave: Make the unchanged TP case
faster, 2020-12-21) we just zero out the AD channels when we run the
testpulse with the same settings again.

The code would work okay if numADCEntries was already set to the correct
value, but it was not initialized, so it was zero.

Let's fix that.

The bug can be experienced with multiple AD channels when the AD
channels hold values larger than the limits and
DC_CheckIfDataWaveHasBorderVals complains.

Reported-by: Lisa Kim