DC_PlaceDataInDAQDataWave: Fix setting POST_SET_EVENT/PRE_SET_EVENT in setEventFlag
setEventFlag is indexed by channel number and not active channel which i
represents.

Bug introduced in 404eca7a (Cleanup of DC_PlaceDataInHardwareDataWave, 2019-05-28).

The reason this did not came up in testing was that we nearly always
only ever test with one headstage where the active channel number and
the active channel index coincide at 0.

And review of the author of this commit failed as well to flag this.
Doing another review of that change in 2019 did not reveal any similiar
issues.