Use correct casing for MIES_Include.ipf file
Our main include file is called MIES_Include.ipf and referenced in
multiple places.

These places did not bother to use the correct casing. As that is in
general not required on Windows.

This used to work quite well, but suddenly CI broke in tools/clean_mies_installation.sh
when doing an installation with an installer package with

 mv: cannot stat 'C:\Users\svc_multipatch/Documents/WaveMetrics/Igor Pro 8 User Files/Igor Procedures/MIES_include.lnk': No such file or directory

We now use the correct casing.