tools/clean_mies_installation.sh: Fix edge case and speed up
With the "installer" installation we currently install MIES for each
version with the installer. But that is not necessary as the installer
installs for all available and supported IP versions.

This also fixes an edge case where the installer fails with return value
2 as it still finds an existing IP9 version when installing for IP8 in
the loop.

We also remove the platform dependent special casing for MINGW and
others. We only support Windows/Mingw nowadays.