Testing/UTF_PA_Tests.ipf: Monkey patch the data to be in IC
When we added the PA tests in 7fac38c8 (Tests: Initial PulseAveraging Tests, 2020-11-04) the sweeps were acquired in VC as that is the default and it did not matter back then.

But now we treat the pulses differently in the failed pulse checks
depending on their clamp mode.

Let's just "tweak" the labnotebook entries to claim it holds IC data to
fix the tests.