HW_WriteDeviceInfo: Fix TTL count for ITC hardware
There are two bugs here.

First of all we don't want to take the minimum of DO and DI counts, as
we use the TTL channels only for output not input.

Second issue is that we have to multiply the DO count by 8 as we can
have 8 "DaEphys TTL channels" on one ITC TTL channel.