Rework device info handling
We used to create device info waves for the locked device and for
ITC1600 pressure devices in P_PrepareITCWaves.

For a future application we want to be able to query device properties
of unlocked and unregistered devices.

We therefore store the device info waves in a separate folder, see
GetDeviceInfoPath(), and name the waves like the device.

We don't offer an upgrade path for the device info waves as those are
generated on fetching the list of devices.