AI_SyncGUIToAmpStorageAndMCCApp: Handle headstages without amplifiers
When "Sync MIES to MCC" is checked in the JSON configuration, we
currently try to sync all settings of all headstages, even the ones
which do not have an amplifier connected.

This just takes a lot of time and also triggers a RTE inside
AI_SendToAmp.

So let's ignore headstages without amplifiers.