Analysis/Labnotebook Browser: Handle multiple experiments with disjunct devices
The labnotebook graphing tool in the analysis browser allows to change
the experiment and device when viewing labnotebook entries.

But in the original implementation from b419fd07 (Add the first version of the Analysis Browser package, 2015-02-25) we did not think about multiple experiments were one has DeviceA and second DeviceB.

And now whenever you change the experiment, you get an assertion from
GetAnalysLBNumericalKeys/GetAnalysLBTextualValues.

So let's handle that case gracefully by either selecting the correct
device if we only have one or complaining to the user.