Fix package folder location on disk
Since c87daee1 (MIES_PackageSettings.ipf: Extract settings folder
creation into separate function, 2021-03-09) we created the log file and
the JSON settings file in the folder

C:\Users\$user\AppData\Roaming\WaveMetrics\Igor Pro 8\Packages

instead of

C:\Users\$user\AppData\Roaming\WaveMetrics\Igor Pro 8\Packages\MIES

.

The reason is that the path contained an additional colon (":") which in
IP lingua means "go down one folder". But that was only the case in IP8,
IP9 has a bug where this is not done. The bug is already reported.

Remove the additional colon so that it works with all IP versions.