Fix and rename DAP_GetHighestActiveHeadstage
The function does currently assume that at least one headstage is in the
given mode.

But if you are requesting the highest headstage for given clamp mode and
no headstage is in that clamp mode, the call never succeeds as NaN is
returned but the caller compares with != and that is always true.

Rework the function to return 0/1 and move it to MIES_DAEphys_GuiState.ipf.

The tests were also adapted. As these tests rely on override functions
in the builtin procedure window, RunTests.pxp was updated as well.