Tests without Hardware: Set interactiveMode to zero in AdditionalExperimentCleanup()
We never want that to be true for the average test case. Test cases
which expect a certain value as default, like DON_WorksWithDefault(),
need to kill the existing variable first.

This also fixes the bug that currently RunTests.pxp pops up dialogs as
the interactiveMode is set to true.

This was not a problem until 9c876ef1 (MIES_BackgroundWatchdog.ipf: Make it self-contained, 2021-03-19) as DoAbortNow() used GetInteractiveModeWrapper. And GetInteractiveModeWrapper() always returned NaN since GetInteractiveMode was made threadsafe in 1bd76a9b (GetInteractiveMode: Mark as threadsafe, 2017-12-21).

So this means DoAbortNow always thought is is operating with
interactiveMode turned off.

This also drastically reduces CI time as RunTests.pxp does not timeout
anymore.