StartZeroMQSockets: Rename and allow forceful restarting
The menu option for this function still says Restart, but since 92d9ce5c
(StartZeroMQMessageHandler: Make it do nothing if it is already running,
2020-03-11) we are not doing anything when it is already running. So
let's add a forceRestart option to do what the menu option says.

While we are at it we also rename the function as it since forever does
also bind zeromq sockets.