MIES_AnalysisFunctions_PatchSeq.ipf: Make shortcut on passed baselineQC work
We are always writing the baselineQC value as headstage dependent data
into the labnotebook. But for PSQ_DAScale/PSQ_Rheobase and PSQ_Ramp we
queried the headstage independent value when shortcutting if we are
done.

This is now fixed.

The bugs were introduced in:
- f3fbfc8f (PSQ_DA_SCALE: Postpone sweep passed decision, 2019-04-13)
- 3bcdaadb (Add Rheobase patch seq analysis function with tests, 2017-12-10)
- 3a7ef027 (PatchSeq Ramp: Add analysis function, 2018-03-20)

The fallout from the bug is slightly worse runtime,
PSQ_EvaluateBaselineProperties would properly return early as it checks
each chunk.