AddTraceToLBGraph: Use a different marker for the first headstage
We are currently using the markers 0-7 for each headstage. This works
well for all headstage except the first one.

The marker 0, of the first headstage, is a plain +.

But for the first sweep this is always completely overlayed by the
y-axis axis and ticks.

By using 39, which is #, this can be avoided.