Packages/doc/dot: Remove subgraph feature
Dotty does not understand that properly. And this results in the default
color of each new node being the subgraph color. So we either have
invisible newly created nodes or set a non-white color to the subgraph
border.

So let's just remove the subgraphs. The files are then also resaved in
dotty.