HW_GetDeviceInfoUnregistered: Rework logic and remove device opening for NI hardware
Introduced in d5d7b011 (HW_GetDeviceInfoUnregistered: Add it,
2021-03-17) but as it turned out opening an NI device also results in it
being resetted and calibrated, but we definitly don't want that here.

This code path is implicitly hit when DAP_GetNIDeviceList needs to
regenerate the list of NI devices. And that at the moment also happens
when exporting a file into NWB via BuildDeviceString and GetAllDevices.

We can't just move the HW_OpenDevice call into the HARDWARE_ITC_DAC
branch as that also determines the hardware type. But we always know the
hardwar type when calling HW_GetDeviceInfoUnregistered so we can just
pass it in.

Reported-By: Lisa Kim
Helped-By: Tim Jarsky