Rename and rework BuildDeviceString
This function is never used for NI hardware. But it always did check for
NI hardware first, which called DAP_GetNIDeviceList and that could
potentially be expensive when the global NI device list is regenerated
as that also includes a call to DAP_UpdateDeviceInfoWaves.

So let's just remove the claimed and never used NI support, thus making
the function faster.

It is also renamed to convey in its name that it is for ITC hardware
now only.