Add GetActiveChannels
We now have a uniform way to answer the question "Which channels were
active when this sweep was acquired.".

This works for all channel types, associated/unassociated channels and
all hardware types.

The situation for TTL channels is also solved here, althought the
TTLmode parameter was required to cover all uses cases.