Add code to allow recovering deleted sweeps
Sweeps can be deleted manually or due to 9c27dcd (SWS_DeleteDataWaves:
Make it faster, 2020-03-26) when sweep rollback is used.

In case the databrowser was open when that happened we still have the 1D
waves lying around. So we can try to reconstruct the sweep and config
waves from the 1D waves and the LBN entries.