Labnotebook: Enforce that the INDEP getters are correctly used
As the last commit showed, we don't have any safety precautions when an
dependent setting is queried as independent.

But we can add a check for CI to enforce that. We don't do that check
outside of CI for speed reasons. The IsConstant check is required for
very old entries which have the same value in all LABNOTEBOOK_LAYER_COUNT
rows.

We can not enforce the other direction, a dependent setting queried as
independent, as the user uses GetLastSetting for that and we don't know
inside GetLastSetting what she wants to do.