Sweepbrowser: Show the sweep number and not the index
Since forever the databrowser and the sweepbrowser differ in the way
they look at their sweeps. The sweepbrowser uses indizes into a list
whereas the databrowser uses the sweep number itself.

The reason for that distinction is that the sweepbrowser can contain
duplicated sweep numbers as the data can be from different experiments.

We now also display the sweep number in the setvar_SweepControl_SweepNo
control for the sweepbrowser.