Dashboard: Display entries for sweeps without analysis function as well
Time has shown that the dashboard is useful for browsing sweeps grouped
by stimset cycle ID.

We therefore don't require a known analysis function being attached
to the stimset anymore.

This means it now works with all sweeps acquired.

Sweeps without stimset cycle ID is still ignored as we can't group by
that propery if it does not exist.