WB_OpenStimulusSetInWaveBuilder: Handle non existing stimsets better
When we try showing a stimulus set we might realize that it does not
exist. Bug introduced in b5bf561f (Add new trace popup for convenient
stimulus set opening, 2021-02-18).

We can not do anything when we are in the databrowser. But we can try
loading the stimulus set when we are in the sweepbrowser.

Close #932