GetWaveSize: Fix crash when invoked from threadsafe function
A bug fixed in WaveInfo fixed in IP9 r37431 results in a crash when
WaveInfo is used in a threadsafe function.

We workaround that issue for older versions by resurrecting the solution
from 996cee17 (Implement GetWaveSize via WaveInfo, 2018-02-16) which was
used before WaveInfo learned the "SizeInBytes" keyword.

The resurrected solution was also cleaned up and some magic numbers
replaced by constants.