NWB_AddDeviceSpecificData: Tweak RemoveUnusedRows to always return a wave
In c5b1ad32 (NWB_AddDeviceSpecificData: Only write the used rows into
the NWB files, 2020-09-24) we added support for dropping unused rows
from labnotebook and tpstorage waves.

Back then we decided that if the wave is empty we don't write it. This
actually does not work currently as we later on want to attach an
attribute for NWBv2 to these datasets.

Thinking about this again, it was also the wrong decision to skip saving
it if it is empty. Especially for the labnotebook, nobody is prepared for
getting less than four waves.

The newly added test ExportIntoNWB revealed these flaws.