MIES_WaveDataFolderGetters.ipf: Make the labnotebook cache waves work in preemptive threads
When running the labnotebook getter functions in preemptive threads the
labnotebook cache wave getters currently assert out as all passed waves
are free waves and WaveModCountWrapper does enforce permanent waves.

Free waves don't have a modification count or date. So we can't really
track if the labnotebook waves have been changed. We could do a WaveCRC
on the whole wave and store that, but this a rather expensive.

But we don't need to check if the wave changed if it can't change
because it is locked.

So let's enforce that the labnotebook waves are locked when running
in a preemptive thread.