Use ASYNC framework to write sweep-by-sweep NWB files
Since forever we have the problem when exporting after each sweep into
the NWB file that this takes a considerable amount of time away from the
ITI and makes IP feel laggy.

Fortunately with the ASYNC framework we have a generic solution for
executing code in a preemptive thread.

The functions are now also a bit reorganized, the documentation at the
top of the MIES_NeuroDataWithoutBorders.ipf file should help with
understanding the logic.

The stimset writing is dropped from the sweep-by-sweep export.