Tests: Fix Abort_ITI_PressAcq_SD/Abort_ITI_PressAcq_MD
These tests abort DAQ during ITI by pressing "Stop data acquisition".

The reentry functions currently check that the last labnotebook entry is
from TP.

This will be broken once we add the "DAQ stop reason" to the labnotebook
in a future commit.

Change the tests to just check for a TP entry in the labnotebook is also
correct and future proof.