ReturnListOfAllStimSets: Rework and rename it
Now all parameters are optional, so the function can be used to query
all stimsets from all channels as well. This also makes
ReturnListOfAllStimSetsFromAllChannelTypes redundant. The new function
is also moved to MIES_WaveBuilder.ipf and renamed to WB_GetStimsetList()
as it fits better there.

CHANNEL_TTL_SEARCH_STRING and CHANNEL_DA_SEARCH_STRING are removed as
parameters to WB_GetStimsetList as they are not needed at all.

DAP_FormatStimSetPopupValue learned to not require
GetSearchStringForChannelType which could then also be removed.