NWB: Ensure that sweep-by-sweep export is finished when required
Since fbbb2f6a (Use ASYNC framework to write sweep-by-sweep NWB files,
2021-05-11) we write the data in sweep-by-sweep export mode via the
ASYNC framework.

But the author of this commit forgot to add a synchronization point,
when closing the NWB file and/or unlocking the device.

This is now done. We do wait for up to 10 minutes for the NWB writing to
finish. This should cover all obscure corner cases, and we also print
out a message every 5s.