MIES_WaveBuilderPanel.ipf: Add GetEpochParameterNames
This function returns a 2D wave which holds all parameter names of all
epoch types.

Due to the unique design of our WaveBuilder we don't have code which
holds that, but the WaveBuilder panel itself together with a naming
scheme for the controls determines the parameters of all epoch types.

Therefore the function WBP_ListControlsPerStimulusType and
WBP_RegenerateEpochParameterNamesCode allow to generate code which then
is used in GetEpochParameterNames.

To ensure that GetEpochParameterNames stays correct we also added a test
for that.