WB_GetStimSet: Allow creating empty stimsets
An empty stimset is one which does not have any row. We already don't
allow data acquisition with it, see last commit, and we also prevent
data acquisition in DAP_CheckSettings.

But for using the new stimset API it is beneficial to be able to create
intermediate stimsets which, although not yet valid, can be made valid
anytime by changing some parameters.