PSQ_CR_DetermineCycles: Fix gain querying
The returned wave from SWS_GetChannelGains holds all active channels.
And we want to have the first active DA channel which is zero, and we
don't index with headstages.

And as we are using timing = GAIN_AFTER_DAQ we also need to divide
through the gains and not multiply with it.

This bug is asymptomatic if you are using Headstage 0 with DA 0 and AD
0.