tools/nextFreePRNumber.sh: Make it work with discussions
Discussion on github share the same number space as issues and PRs.
Unfortunately we can not just query the list of discussions using the
REST API as there only exists a GraphQL API and that needs always
authentication.

So let's just use dump curl to find the first discussion ID which return
a HTTP status code different than 200.