MIES_Async.ipf: Add support for running without threading
For debug purposes it is nice if the async framework would still work
even when THREADING_DISABLED is defined.

This commit adds this support. The idea is that ASYNC_Execute already
executes the workload using ASYNC_Run_Worker and then stores the result
in a local DFREF wave called serialExecutionBuffer.

ASYNC_ThreadReadOut then looks in that wave instead of
ThreadGroupGetDFR.

Some tests which check that aborting with running workload are disabled
as they don't pass anymore.