MIES_NeuroDataWithoutBorders.ipf: Write epoch information into NWBv2 files
We now write epoch information into NWBv2 files then exporting whole
experiments into NWBv2.

The updating of fileIDExport is unfortunately necessary as
WriteSingleChannel uses a custom XOP for writing the timeseries
datasets. And therefore we have to close and reopen the HDF5 file which
results in a new ID.

This also the reason we don't export epoch information on sweep-by-sweep
export.