SC_SpikePositionQC: Don't bug out on debug string generation
spikePositionsPerPulseNum was indexed incorrectly as we need to index
with the pulse numbers and not the headstage index.

By itself this would not result in a user visible issue. But as
RemovePrefix did query GetRTError() without clearing it first, this
aborted the current sweep.

Bug introduced in f2292ce2 (Enhance MSQ_SpikeControl, 2021-02-02).