MIES_AnalysisBrowser.ipf: Allow resaving an existing NWBv1 as NWBv2
For users who have lost the pxp but still have the NWBv1 file, we are
now offering a way to convert from NWBv1 to NWBv2.

We convert sweep data, stimulus sets, labnotebooks and stored test
pulses and do that by first loading the data, then moving it into the
usual location for data acquisition and then just save as NWBv2.

All loaded NWBv1 files from the selected folder in the analysis browser
are converted.