PSQ_GetLastPassingLongRHSweep: Return the correct sweep
Since the initial implementation of the new PSQ_DaScale operation mode
in 0ef300da (PSQ_DaScale: Add new operation mode, 2018-02-15) the logic
to find the reference rheobase sweep was flawed.

It did not check that the Rheboase Set QC entry did pass, it only
checked that it existed, so failing was also okay. And it just returned
the sweep with a Set QC entry. But actually we want to have a spiking
sweep from an SCI which has passing Rheobase Set QC and the correct
duration.

This is now fixed. In addition the duration is now a parameter which
allows to test the function. Formerly this code was never exercised in
CI testing.