Analysis Functions: Remove PRE_SWEEP_EVENT and repurpose as PRE_SWEEP_CONFIG_EVENT
Experience has shown that the current PRE_SWEEP_EVENT, which happens
after the sweep is configured and immediately before DAQ starts, is off
very limited use.

In all PSQ/MSQ analysis functions it is never used.

The reason is that as it is called after the sweep is configured, the
analysis function can not change anything relevant anymore.

Instead of now adding a new event which is called before the sweep is
configured, the existing one PRE_SWEEP_EVENT is repurposed as
PRE_SWEEP_CONFIG_EVENT.