SetControlInEvent: Be backwards compatible with existing "Pre Sweep" entries
This analysis function supports setting GUI controls in various analysis
function events.

This means we have serialized the "Pre Sweep" name into analysis
function parameters stored in user stimulus sets.

We therefore now translate the old "Pre Sweep" event name to the new name
and complain to the user.

Also add a test with new and old names to verify that it works.