Databrowser/Sweepbrowser: Allow selecting sweeps and set sweep/cycle counts
We now allow selecting sweeps by their stimset and set sweep count/set
cycle count.

In order to avoid cluttering the GUI too much, we only show stimsets
which cycled at least once.

The new names in the popupmenu are `$stimset ($setSweepCount)` and
`$stimset (C$setCycleCount)` as they have to be unique for the selection
code to work.