PSQ Analysis functions: Rework decision logic for post pulse baseline chunks
Since forever we have PSQ_EvaluateBaselineProperties return NaN for both
not-enough-data and and a failing post pulse baseline chunk.

Assume we have the following chunks (passing chunks have a *):

| pre pulse* | PUL- | SE | ignored | post pulse I | post pulse II* |

Assume we are now called at the very end of DAQ with all data acquired
and all post pulse chunks not yet evaluated. This can happen for example
on a slow computer.

If "post pulse I" fails, we would immediately stop and mark the whole
baseline as failure. But if we would look into "post pulse II" we would
see that it passes and the full baseline would pass as only one post
pulse chunk has to pass.

The fix is to not reuse the same constant for not-enough-data (NaN) and
post pulse baseline chunk failed (NaN). The latter is now PSQ_BL_FAILED.

In addition the check for not enough data and the last chunk is
tightened as we now only fail that chunk if we have acquired all the
data.

Fallout:
- Only present on slow computers
- The first failing post pulse baseline makes the whole baseline fail
  wheras it still could pass if later chunks were to be looked at.

Found with UTF tracing.