Testpulse settings: Rework them completely
Since forever we had multiple globals for TP parameters (as entered by
the user into the GUI) and calculated TP values.

The reasoning was that using globals is faster than querying the GUI
(for entered) values.

But having a lot of globals makes the code more difficult to read/write
and also clutters the datafolders.

We now create the following waves to handle that situation better:
- GetTPSettings(): Holds all TP settings from the Data Acquisition tab as
  entered by the user. It also holds analysis relevant TP entries from
  the Settings tab like tolerance and buffer size
- GetTPSettingsCalculated(): Holds all calculated/derived TP settings.
  This makes TP_GetTestPulseLengthInPoints/GetTestpulseLengthInPoints
  obsolete which were brittle and difficult to use. It is only valid
  during DAQ/TP.
- GetTPSettingsLabnotebook/GetTPSettingsLabnotebookKeyWave: This stores
  the TP settings which were used for DAQ/TP for later LBN writing. The
  approach is the same as for the sweep settings waves. The reason this
  is required is that the user might change the GUI entry during DAQ/TP
  but the saved settings should be the used settings and not the current
  ones. This issue was formerly solved by using the now removed globals.

These waves are then update/initialized when required.

This also fixes a bug where the wrong testpulse duration was fed into
TP_TSAnalysis. This error was in SCOPE_UpdateOscilloscopeData were the
plain duration was used. And that duration is in points for the test
pulse sampling interval, which is different to the DAQ sampling interval
when TTL channels are active or sampling interval multiplier was used.
Bug is only present with TP during DAQ.

Close: #1015