Rework TP results handling completely
Up to now we had multiple 1D waves (GetBaselineAverage,
GetSSResistanceWave, GetInstResistanceWave) for storing the TP analysis
results. And we also had these waves duplicated for the buffered values
for averaging (GetGetBaselineBuffer, GetInstantaneousBuffer,
GetSteadyStateBuffer).

This makes it quite laborious to add new values.

We now therefore switch to two waves (GetTPResults and
GetTPResultsBuffer) for storing the immediate TP result and the buffered
values.

TP_CalculateAverage is also reworked to be easier to understand and now
tested as well.