Allow browsing TPStorage/Labnotebook waves in Databrowser/SweepBrowser
Up to now the user could only browser labnotebook waves in the
databrowser and had to use the separate Labnotebook browser when loading
experiments via the analysis browser.

On the other hand TPStorage waves could only be browsed via the
analysis browser and not live in the data browser.

This commit introduces the functionality to browse the
TPStorage/Labnotebook waves from both the Databrowser and the
SweepBrowser.

This is done by allowing to select the current experiment and device in
the GUI.

Code relating to this is now also placed into the new file
MIES_LogbookViewer.ipf.