GetAnalysLBN*: Allow non-existent labnotebooks
In edge cases like only stimsets in the experiment/NWB file it is totally
expected to not have a labnotebook.

And as the upcoming labnotebook browser could still be used by the user
we need to fail gracefully in this case.

Add assertions for existing callers which expect a wave.