SB_SweepBrowserWindowHook/DB_WindowHook: Don't kill the window
It is bad style according to the documentation to actually kill the
window in the kill event.

And, much more severe, it does prevent other hooks being called on the
now killed window.

In our case the TUD_RemoveUserDataWave hook was not called, thus making
the TUD data lying stale around.

Fix that by not killing the window. Our KillOrMoveToTrash call handles
the case that the waves we want to kill are moved away regardless if
they are displayed or not.