Analysis function dashboard: Fix the failure messages and make them more verbose
The analysis function dashboard displays the analysis function result for
each stimulus set cycle ID (SCI).

But we want to have the failure result of each single sweep. This was not
implemented instead e.g. AD_GetBaselineFailMsg look at the whole SCI. And
this triggered an assertion with
Chat-IRES-Cre-neo;Ai14-582721.05.04.01.nwb as from three sweeps the middle
one was failing but the last one passed.

Investigating this assertion revealed that this approach is fundamentally
broken as we need to query each sweep failure mode and then, depending on
the analysis function, other things as well for the whole SCI.

The new AD_GetPerSweepFailMessage does that.