PEXT_SplitToSubMenus: Calculate the number of submenu entries dynamically
The hardcoded limit of 30 does not work when there are a large number of
labnotebook keys. A larger hardcoded limit would not work with fewer
entries.

So we now calculate the number of entries by calculating an upper limit
per menu. These upper limits are not reached for some methods as those
can break on e.g. letter boundaries for nicer display.