DAP_UpdateDaEphysStimulusSetPopups: Update also the clamp mode all controls
In 98082cd0 (DA_EPhys: Add controls for changing all DAC of a given
clamp mode., 2016-12-08) we added clamp mode specific all controls for
the DACs.

But it was forgotten that the "MenuExp" user data needs to be updated
for these as well. Fix that.

The loop is now split in two as there are no clamp mode specific all
controls for TTL channels.