DAP_UpdateDaEphysStimulusSetPopups: Handle adding stimsets better
Recipe from Tim Jarsky:
Set up Wave_DA_AllIClamp and IndexEnd_DA_AllIClamp, and check
check_DA_applyOnModeSwitch. Now create stimset with a name that causes it
to slot in above the stimsets selected for indexing. e.g., If start index
stimset is named B, create a new stimset named A. Switch clamp modes and
the start index popup-menu **on the switched headstage** is offset by
one."

The reason for this offset is that IP and our GUI state wave still have
the old indizes of the PopupMenu's stored, as both don't know that the
list of stimsets changed. But that old index is no off-by-one, or even
more if the user added multipe stimsets.

Removing stimsets does not have that problem as DAP_PopupMenuWaveNameList
and DAP_RestorePopupMenuSelection handle this already. That approach is
quite expensive though as it uses PGC_SetAndActivateControl for every
control. It also requires to store the lists before changing the stimset
count and applying them again.

This is quite inconvenient. Tests show that we actually don't need to
store anything before. The only thing we need to do is to get the correct
index and store it in IP and the GUI state.

Broken since forever.