Sweepformula did not process comments with # correctly
The issue was that the "vs" token for plotting was processed before the
comments were processed in the state based parsing. Since the "vs" token is
parsed with a regex the parsing can not be simply swapped.

Fix: Before the formula is parsed it is preprocessed where based on a per line
basis the comments after # are removed.
The same step is used to unify the line endings to CR by replacing LF with CR.
(CR LF -> CR CR adds an empty line, but the state based parser ignores empty
 lines)

The comment handling in the state based parser was removed.

As testing calls the Formula Parser function directly some additional code was
introduced to execute the preprocessing then there.