Revisited all occurences of ClearRTError prior try constructs
In some cases we run code that may result in an RTE that we want to catch.
Our previous approach was that we enclose this part in a
try / catch / endtry block, where a critical RTE is caught through
an AbortOnRTE.
To prevent catching pending RTEs instead of the specific targeted RTE
we cleared any RTE condition with ClearRTError() directly after the try.

This poses a problem: Any pending RTE is silently cleared by the code which
generally hid all pending RTEs in the code that appeared before one of the
try clauses where executed. While for non-threadsafe code this still might
be found by executing with Debug-On-Error, with threadsafe code such conditions
got invisible.

The fix is to check the RTE code before clearing it as preparation for
critical code. So we can assert out, if there is a pending RTE.